Split out new experimental method mt_to_table_of_ndarray
