+++ hail/python/hail/experimental/__init__.py
@@ -22,6 +22,7 @@ from .loop import loop
 from .time import strftime, strptime
 from .pca import pc_project
 from . import dnd
+from .table_ndarray_utils import mt_to_table_of_ndarray
 
 __all__ = ['ld_score',
            'ld_score_regression',
@@ -61,4 +62,5 @@ __all__ = ['ld_score',
            'strptime',
            'strftime',
            'pc_project',
-           'dnd']
+           'dnd',
+           'mt_to_table_of_ndarray']
